requirements.txt
google-cloud-aiplatform>=1.38.0
google-auth>=2.17.0
pandas>=2.0.0